// React/app/GerAdv_TS/AreasJustica/Interfaces/interface.AreasJustica.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
// copyright © 2000-2025 Menphis - Sistemas Inteligentes
// This file is part of the Source Genesys project                     
'use client';

export interface IAreasJustica {
// 202501251
    id: number;
 
	area: number,
	justica: number,
}

export interface IAreasJusticaFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IAreasJusticaIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}